+++ web/nearfs/cid.spec.js
@@ -1,0 +1,12 @@
+import {CID2String} from './cid.js';
+
+describe('cid', () => {
+    it('should convert cid to string', () => {
+        const cidbytes = new Uint8Array([0x01,0x55,0x12,0x20,0xe1,0xe5,0xd7,0xc7,0x3c,0xff,0x1c,0x8c,0xf7,0x1c,0xac,0xcc,0xf0,0x95,0x22,0x8f,0xa2,0x5e,0x22,0x0e,0xe9,0x43,0x2a,0x13,0x6d,0xfe,0x11,0x0a,0xf0,0x30,0xb8,0x39]);
+        expect(CID2String(cidbytes)).equal('bafkreihb4xl4oph7dsgpohfmztyjkiupujpcedxjimvbg3p6cefpamfyhe');
+    });
+    it('should convert another cid to string', () => {
+        const cidbytes = new Uint8Array([0x01,0x55,0x12,0x20,0x6e,0x6d,0x8c,0x81,0x38,0x76,0x17,0xab,0x22,0x28,0x0e,0x83,0xdb,0xbe,0x46,0x73,0x04,0x9c,0x9c,0x78,0x37,0xe9,0x4f,0xda,0xec,0x7e,0x51,0xf0,0xf6,0x28,0x33,0x4a]);
+        expect(CID2String(cidbytes)).equal('bafkreidonwgicodwc6vsekaoqpn34rttasojy6bx5fh5v3d6khypmkbtji');
+    });
+});
